All tests are now in working state. application.py has been moved to api/
