+++ api/tests/test_auth_functions.py
@@ -12,7 +12,8 @@ sys.path.append(normpath(join(SCRIPT_DIR, PACKAGE_PARENT)))
 
 from manage import seed, remove_seed
 seed()
-from api import app, db
+from app import app
+from db import db
 from models import Users, User_affiliations, Organizations
 from functions.auth_functions import (
     is_super_admin,
