+++ api/tests/test_groups_resolver.py
@@ -10,7 +10,8 @@ from unittest import TestCase
 from manage import seed, remove_seed
 
 seed()
-from api import app, db
+from app import app
+from db import db
 from models import Sectors, Groups
 from queries import schema
 from backend.security_check import SecurityAnalysisBackend
