+++ api/tests/test_organizations_resolver.py
@@ -11,7 +11,8 @@ from unittest import TestCase
 from manage import seed, remove_seed
 
 seed()
-from api import app, db
+from app import app
+from db import db
 from models import Groups, Organizations
 from queries import schema
 from backend.security_check import SecurityAnalysisBackend
