+++ api/tests/test_scans_resolver.py
@@ -12,7 +12,8 @@ from unittest import TestCase
 from manage import seed, remove_seed
 
 seed()
-from api import app, db
+from app import app
+from db import db
 from models import Scans, Domains, Users
 from queries import schema
 from backend.security_check import SecurityAnalysisBackend
