+++ api/tests/test_that_migrations_have_a_single_revision_head.py
@@ -6,7 +6,7 @@ from alembic.script import ScriptDirectory
 # https://blog.jerrycodes.com/multiple-heads-in-alembic-migrations/
 def test_only_single_head_revision_in_migrations():
     config = Config()
-    config.set_main_option("script_location", "app:migrations")
+    config.set_main_option("script_location", "api:migrations")
     script = ScriptDirectory.from_config(config)
 
     # This will raise if there are multiple heads
